move ConfirmationModal out of Wildcard
