+++ client/web/src/enterprise/insights/components/insights-view-grid/components/ConfirmationModal/index.ts
