+++ client/wildcard/src/components/index.ts
@@ -43,7 +43,6 @@ export {
     createRectangle,
 } from './Popover'
 export { Collapse, CollapseHeader, CollapsePanel } from './Collapse'
-export { ConfirmationModal } from './ConfirmationModal'
 
 /**
  * Type Exports
@@ -61,4 +60,3 @@ export type { IconProps } from './Icon'
 export type { Point } from './Popover'
 export type { BadgeProps, ProductStatusType, BaseProductStatusBadgeProps } from './Badge'
 export type { AnchorLinkProps } from './Link'
-export type { ConfirmationModalProps } from './ConfirmationModal'
